+++ CHANGELOG.md
@@ -3,12 +3,32 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
-and this project adheres to [Semantic Versioning](http://semver.org/).
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.1] Released
+## [1.0.2] - 2025-09-24
+
+### Added
+
+- Added test cases
+
+### Fixed
+
+- Fixed a typo in Assert-Exists where the With alias was called Width
+- Fixed an issue with multi level array validation tests[*].users[1] was failing to properly validate before
+
+### Changed
+
+- Updated README
+- Updated CHANGELOG
+
+## [1.0.1] - 2025-09-23
+
+### Fixed
 
 - Fixing issue with powershell 5.1 compiling the c# code.
 
-## [1.0.0] Released
+## [1.0.0] - 2025-09-23
+
+### Added
 
 - Initial release